Extract license threat so it can be used more places
